Une version qui donne une jolie simulation

+++ observer.rs
@@ -78,7 +78,7 @@ pub impl DiskWriter {
 
     fn new_filename(&mut self) -> ~str {
         self.curr_snap+=1;
-        fmt!("./%s-%u.vtk", self.fname, self.curr_snap-1)
+        fmt!("./%s-%03u.vtk", self.fname, self.curr_snap-1)
     }
 }
 
@@ -103,7 +103,7 @@ impl Observer for DiskWriter {
         // Données sur les nœuds
         writer.write_str(fmt!("POINT_DATA %u\nSCALARS %s float\nLOOKUP_TABLE default\n", crowd.size(), self.fname));
         for crowd.cells.each |&cell| {
-            writer.write_str(fmt!("%u\n", cell.age));
+            writer.write_str(fmt!("%f\n", cell.age as float));
         }
     }
